+++ src/search_core.h
@@ -11,7 +11,7 @@ struct QueryFilters {
     std::string patient_id;
     std::string barcode;
     std::string patient_name;
-    std::string bed_code;
+    std::string patient_no;
     std::string oper_no;
     std::string start_date;
     std::string end_date;
@@ -35,6 +35,7 @@ struct ReportRow {
     std::string bed_code;
     std::string patient_type;
     std::string requester;
+    std::string reviewer;
     std::string group_name;
     std::string conf;
     std::string chk_flag;